+++ Snake/SnekManager.h
@@ -5,10 +5,10 @@
 
 class SnekManager {
 public:
-	SnekManager(Draw*, uint32_t, uint32_t);
+	SnekManager(Draw*, short, short);
 	Draw* GetDraw() { return m_draw; };
-	unsigned int width = 0;
-	unsigned int height = 0;
+	short width = 0;
+	short height = 0;
 private:
 	Draw* m_draw;
 };
